% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{View misuvi data dictionary}
\usage{
dictionary()
}
\description{
This package provides a tidy version of the MISUVI data sets. This function returns a data frame of the cleaned variable names and the original variable names.
}
