% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation.R
\name{documentation}
\alias{documentation}
\title{View the MI-SUVI Documentation}
\usage{
documentation()
}
\value{
NULL, opens a web browser of the technical documentation.
}
\description{
This function opens up a PDF of the MI-SUVI documentation in your web browser.
}
\examples{
documentation()

}
