% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regional_groupings}
\alias{regional_groupings}
\title{Regional Groupings for Michigan}
\format{
`regional_groupings`
A data frame with 83 rows and 10 columns.
}
\source{
Downloaded from the State of Michigan's website: https://www.michigan.gov/opioids/-/media/Project/Websites/opioids/documents/Public-Use-Dataset-County-Population-and-Regional-Groupings.xlsx?rev=6ac0d413e9164876b1bcd32acfc30f0c
}
\usage{
regional_groupings
}
\description{
This dataset provides cross walks between Michigan Counties and PIHPs,
regions, urbanicity groupings, emergency preparedness regionals, and
Michigan State Police Districts.
}
\keyword{datasets}
