% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{dictionary}
\alias{dictionary}
\title{View misuvi data dictionary}
\usage{
dictionary()
}
\value{
A data.frame of the abbreviated and full names of the MI-SUVI variables.
}
\description{
This package provides a tidy version of the MI-SUVI data sets. This function returns a data frame of the cleaned variable names and the original variable names.
}
\examples{
dictionary()

}
